+++ help/help.go
@@ -3,28 +3,28 @@ package help
 import (
 	"strings"
 
+	"clx/constants/margins"
+	"github.com/muesli/reflow/indent"
+
 	"clx/app"
 	"clx/constants/unicode"
 	"clx/info"
-	"clx/screen"
 	"github.com/charmbracelet/lipgloss"
 )
 
 const (
-	newLine = "\n"
-	newPar  = "\n\n"
+	newPar = "\n\n"
 )
 
 func GetHelpScreen(enableNerdFonts bool) string {
-	screenWidth := screen.GetTerminalWidth()
 	textWidth := 70
 
 	var sb strings.Builder
 
-	sb.WriteString(unicode.ZeroWidthSpace + lipgloss.PlaceHorizontal(screenWidth, lipgloss.Center, "") + newPar)
-	sb.WriteString(unicode.ZeroWidthSpace + lipgloss.PlaceHorizontal(screenWidth, lipgloss.Center, info.GetText(textWidth, enableNerdFonts)) + newPar)
+	sb.WriteString(unicode.ZeroWidthSpace + newPar)
+	sb.WriteString(unicode.ZeroWidthSpace + indent.String(info.GetText(textWidth, enableNerdFonts), margins.MainViewLeftMargin) + newPar)
 
-	sb.WriteString(unicode.ZeroWidthSpace + lipgloss.PlaceHorizontal(screenWidth, lipgloss.Center, lipgloss.NewStyle().Faint(true).Render("press q to return • github.com/bensadeh/circumflex • version "+app.Version)) + newPar)
+	sb.WriteString(unicode.ZeroWidthSpace + indent.String(lipgloss.NewStyle().Faint(true).Render("press q to return • github.com/bensadeh/circumflex • version "+app.Version), margins.MainViewLeftMargin) + newPar)
 
 	return sb.String()
 }